Add React-router & Created some dir
