+++ health-on-time/src/Pages/Home.js
@@ -1,0 +1,12 @@
+import React from "react";
+import LandingPage from "../Components/landingPage";
+
+const Home = () => {
+  return (
+    <div>
+      <LandingPage />
+    </div>
+  );
+};
+
+export default Home;
