+++ health-on-time/src/Pages/MedicineCabinet.js
@@ -1,0 +1,13 @@
+import React from "react";
+import SearchMedication from "../Components/SearchMedication";
+
+const MedicineCabinet = () => {
+  return (
+    <div>
+      <h1>MedicineCabinet</h1>
+      <SearchMedication />
+    </div>
+  );
+};
+
+export default MedicineCabinet;
