+++ health-on-time/src/components/landingPage.js
@@ -1,15 +1,13 @@
-import "../App.css";
+import "../Styles/App.css";
 import React from "react";
 import { Jumbotron, Button, Card, CardDeck } from "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
-import SearchMedication from "./SearchMedication";
 
 const LaningPage = () => {
   return (
     <div className="App">
       <header className="App-header"></header>
       <main>
-        <SearchMedication />
         <img alt="hero" className="hero" src="/hero.jpg" />
         <Jumbotron className="jumbtron">
           <p>
